Move deploy folder inside the dashboard folder
